fix: reliable world-UI layering and raycast at large world coordinates
The world-UI controls and the video quad were submitted on the same order(0)
layer and sorted by upload distance, so at coordinates like 10000+ the float
distance precision could no longer separate the UI's ~0.005 z-layer from the
video and controls flickered in and out.

- Submit all world-UI quads at order(1) so controls paint after the order(0)
  video layer regardless of distance sorting
- Render the UI in the screen's local frame (translate to the camera-relative
  plane centre once, then submit small local-coordinate vertices) so the UI's
  z-layers stay exact at large coordinates
- Run the screen raycast in double precision inside the local frame and carry
  the plane centre as double, keeping hover/click accurate far from spawn
- Centre the UI plane on the video quad's real pivot (anchor + world-Y half
  height) so pitch/roll screens no longer detach the controls from the video

+++ mcedia-mtv/versions/26.2/src/client/java/top/tobyprime/mcedia_mtv/client/worldui/MtvWorldUiInputHook.java
@@ -3,7 +3,7 @@ package top.tobyprime.mcedia_mtv.client.worldui;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.minecraft.client.Minecraft;
 import net.minecraft.world.phys.HitResult;
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 import top.tobyprime.mcedia_mtv.client.channel.ClientChannelPlaybackManager;
 import top.tobyprime.mcedia_mtv.client.channel.worldui.WorldUiCapabilityState;
 import top.tobyprime.mcedia_mtv.client.channel.worldui.WorldUiControlSender;
@@ -87,7 +87,7 @@ if (selection != null && MtvWorldUiRenderer.presentation().isPlaylistExpanded())
                 .filter(EntityPlayerHandle.WorldUiScreen::powered)
                 .filter(s -> s.channelId() != null && !s.channelId().isBlank())
                 .toList();
-        var hit = WorldUiScreenRaycast.select(new Vector3f((float) origin.x, (float) origin.y, (float) origin.z), new Vector3f(camera.forwardVector()), screens.stream().map(EntityPlayerHandle.WorldUiScreen::plane).toList()).orElse(null);
+        var hit = WorldUiScreenRaycast.select(new Vector3d(origin.x, origin.y, origin.z), new Vector3d(camera.forwardVector()), screens.stream().map(EntityPlayerHandle.WorldUiScreen::plane).toList()).orElse(null);
         if (hit == null || blockedByBlock(client, origin, hit.distance())) return null;
         for (var screen : screens) if (screen.plane() == hit.screen()) {
             var snapshot = ClientChannelPlaybackManager.getInstance().snapshot(screen.channelId());
